Make scaling up to >int.Max site easier by moving to using a number of ubiquitously-used datatypes. Difftest works.

+++ Code/lb/SimulationState.h
@@ -9,8 +9,8 @@ namespace hemelb
   {
     struct SimulationState
     {
-        long CycleId;
-        unsigned int TimeStep;
+        unsigned long CycleId;
+        unsigned long TimeStep;
         double IntraCycleTime;
         int IsTerminating;
         int DoRendering;
